refactor: decompose Header component following KISS principles
- Split monolithic Header.tsx (271 lines → 27 lines) into focused components
- Extract SearchBar: dedicated search logic with debouncing and suggestions
- Extract Navigation: home/catalog links and page title display
- Extract UserMenu: authentication-aware cart/wishlist/profile actions
- Improve maintainability: each component has single responsibility
- Enhance testability: components can be tested in isolation
- Reduce complexity: eliminate 8+ mixed responsibilities in single file

Addresses KISS principle violations identified in code quality audit.
